move the library code from `run.rs` to `nu-std`
