+++ crates/nu-std/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+authors = ["The Nushell Project Developers"]
+description = "The standard library of Nushell"
+repository = "https://github.com/nushell/nushell/tree/main/crates/nu-std"
+edition = "2021"
+license = "MIT"
+name = "nu-std"
+version = "0.78.1"
+
+[dependencies]
